+++ examples/data-sources/hopsworksai_azure_user_assigned_identity_permissions/data-source.tf
@@ -1,0 +1,9 @@
+# default permissions
+data "hopsworksai_azure_user_assigned_identity_permissions" "permissions" {
+
+}
+
+# disable backup permissions
+data "hopsworksai_azure_user_assigned_identity_permissions" "permissions" {
+  enable_backup = false
+}