+++ hopsworksai/provider.go
@@ -41,8 +41,9 @@ func Provider(version string) func() *schema.Provider {
 				},
 			},
 			DataSourcesMap: map[string]*schema.Resource{
-				"hopsworksai_clusters":                    dataSourceClusters(),
-				"hopsworksai_aws_instance_profile_policy": dataSourceAWSInstanceProfilePolicy(),
+				"hopsworksai_clusters":                                 dataSourceClusters(),
+				"hopsworksai_aws_instance_profile_policy":              dataSourceAWSInstanceProfilePolicy(),
+				"hopsworksai_azure_user_assigned_identity_permissions": dataSourceAzureUserAssignedIdentityPermissions(),
 			},
 			ResourcesMap: map[string]*schema.Resource{
 				"hopsworksai_cluster": clusterResource(),