+++ CHANGELOG.md
@@ -6,3 +6,4 @@ FEATURES:
 * **New Resource**: `hopsworksai_cluster`
 * **New Data Source**: `hopsworksai_clusters`
 * **New Data Source**: `hopsworksai_aws_instance_profile_policy`
+* **New Data Source**: `hopsworksai_azure_user_assigned_identity_permissions`